requirements.txt
streamlit
pandas
openai
python-dotenv
numpy
jupyter
ipykernel
pyarrow
supabase
rapidfuzz
